+++ lib/king_cache.dart
@@ -13,19 +13,25 @@ import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:http/http.dart';
 import 'package:in_app_update/in_app_update.dart';
+import 'package:intl/intl.dart';
 import 'package:local_auth/local_auth.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:url_launcher/url_launcher.dart';
 
+import 'src/cache_manager.dart';
 import 'src/cache_via_rest.dart';
 import 'src/log_methods.dart';
-import 'src/network_request.dart';
 
+export 'src/cache_manager.dart';
+
+part 'src/custom_widget.dart';
 part 'src/enums.dart';
 part 'src/global_methods.dart';
+part 'src/global_extensions.dart';
 part 'src/king_cache.dart';
+part 'src/network_request.dart';
 part 'src/response_model.dart';
-part 'src/custom_widget.dart';